refactor(view:matter): moved animator directory to character
アニメーター等のユーティリティー系のディレクトリが増えると
構造が見えにくいので、Actorに移動した。
