+++ src/view/matter/actors/character/animator/characterFadeoutAnimator.ts
