+++ src/view/matter/actors/character/animator/characterPopoutAnimator.ts
