+++ src/view/matter/controllers/character/popupCharacterController.ts
@@ -2,9 +2,9 @@
 import { Controller } from 'src/view/matter/types/controller';
 import P5Types from 'p5';
 import { World } from 'src/view/matter/types/world';
-import { CharacterPopoutAnimator } from 'src/view/matter/animator/character/characterPopoutAnimator';
+import { CharacterPopoutAnimator } from 'src/view/matter/actors/character/animator/characterPopoutAnimator';
 import { CharacterParams } from 'src/view/matter/actors/character/characterParams';
-import { CharacterFadeoutAnimator } from 'src/view/matter/animator/character/characterFadeoutAnimator';
+import { CharacterFadeoutAnimator } from 'src/view/matter/actors/character/animator/characterFadeoutAnimator';
 import { Animator } from 'src/view/matter/types/animator';
 
 const AnimatorKey = {
